2 suits
Added the ability to play with two decks. There's a slight bug where
it'll crash sometimes if two cards of the same value are drawn, but it
only happens so often and I'm able to exactly why.

Also: test.cc -> main.cc, because that's it's not really a test anymore.